requirements.txt
asgiref==3.3.1
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3.1
Django==3.1.5
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
google-api-core==1.25.0
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.8
Pillow==8.1.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyparsing==2.4.7
python-decouple==3.4
pytz==2020.5
requests==2.25.1
rsa==4.7
ruamel.yaml==0.16.12
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
